requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
docker==6.1.3
gitdb==4.0.10
GitPython==3.1.37
idna==3.4
packaging==23.1
requests==2.31.0
smmap==5.0.1
urllib3==2.0.5
websocket-client==1.6.3
